Add support for arrays in packets

+++ Network/Blocks.Net.PacketSourceGenerator/Attributes/Packet.cs
@@ -3,4 +3,4 @@
 namespace Blocks.Net.PacketSourceGenerator.Attributes;
 
 [AttributeUsage(AttributeTargets.Class | AttributeTargets.Struct)]
-public class Packet(byte id, bool clientBound = true, string state = "Play") : Attribute;
+public class Packet(int id, bool clientBound = true, string state = "Play") : Attribute;
